Fix for failing tests on Sierra dashboard: Initializing mOutputSTL variable. It wasn't getting initialized and depending on the OS it was running on it was potentially getting non-zero values. Refactored an ESP header for clarity.
